+++ examples/vhdl/pipeline/pipelined_adder.toml
@@ -14,5 +14,5 @@ uut = 'uut'
 
 [flow.openroad]
 platform = "sky130hd"
-clocks.main_clock.freq = "500 MHz"
+clocks.main_clock.freq = "200 MHz"
 core_utilization = 45
