+++ src/__init__.py
@@ -3,15 +3,19 @@ from .cocotb import Cocotb
 from .design import Design
 from .flow import FPGA, Flow, SimFlow, SynthFlow
 from .tool import Tool
+from .flow_runner import FlowRunner, DefaultRunner, Dse
 from .version import __version__
 
 __all__ = [
     "__version__",
     "cli",
     "Cocotb",
+    "DefaultRunner",
     "design",
     "Design",
+    "Dse",
     "flow_runner",
+    "FlowRunner",
     "flows",
     "Flow",
     "FPGA",
