+++ src/flows/dc/templates/run.tcl
@@ -9,7 +9,7 @@ set adk_dir                       {{adk.path}}
 set dc_target_libraries           { {{- adk.target_libraries|join(' ') -}} }
 set mw_ref_libs                   "$adk_dir/{{- adk.milkeyway_reference_libraries|join(' $adk_dir/') -}}"
 set mw_tf                         $adk_dir/{{adk.milkeyway_technology_file}}
-set dc_reports_dir                {{reports_dir}}
+set dc_reports_dir                {{settings.reports_dir}}
 set dc_results_dir                {{settings.outputs_dir}}
 
 file mkdir ${dc_reports_dir}
