+++ src/flows/openroad/templates/filler.tcl
@@ -3,4 +3,4 @@ set_propagated_clock [all_clocks]
 filler_placement {{platform.fill_cells|join(" ")|embrace}}
 check_placement
 
-{{write_checkpoint(step_id)}}
+{{ write_checkpoint(step) }}
