+++ src/flows/openroad/templates/macros.tcl.j2
@@ -11,8 +11,8 @@ puts "{{ line_char * line_len }}"
 {% endif %}
 {% endmacro %}
 
-{% macro preamble(step, index=0, total=1) %}
-{% set msg = (" Starting %s (%d/%d) "|format(step, index, total)) if index else (" Starting %s "|format(step)) %}
+{% macro preamble(step, step_index) %}
+{% set msg = (" Starting %s (%d/%d) "|format(step, step_index + 1, total_steps)) %}
 {{ banner(msg) }}
 {% endmacro %}
 
@@ -20,32 +20,38 @@ puts "{{ line_char * line_len }}"
 {{ banner("( %s done )"|format(step), line_char="-", top_line=false, bottom_line=false) }}
 {% endmacro %}
 
-{% macro load_checkpoint(db_step_id, sdc_step_id=none) %}
-{% if db_step_id %}
-load_db {{results_dir}}/{{db_step_id}}.odb
-{% endif %}
-{% if sdc_step_id %}
-load_sdc {{results_dir}}/{{sdc_step_id}}.sdc
+{% macro load_checkpoint(step, sdc_from=none) %}
+{% set prev_step_id = get_prev_step_id(step) %}
+{% if prev_step_id and should_load_checkpoint(step) %}
+read_libraries
+read_db {{settings.checkpoints_dir}}/{{prev_step_id}}.odb
+{% if sdc_from %}
+{% set sdc_step_id = get_step_id(sdc_from) %}
+read_sdc {{settings.results_dir}}/{{sdc_step_id}}.sdc
 {% endif %}
 {% if platform.derate_tcl %}
 source {{platform.derate_tcl}}
 {% endif %}
+{% if platform.setrc_tcl %}
 source {{platform.setrc_tcl}}
+{% endif %}
+{% endif %}
 {% endmacro %}
 
-{% macro write_checkpoint(step_id, db=true, sdc=false, def=false, verilog=false) %}
-{% if should_write_checkpoint(step_id) %}
+{% macro write_checkpoint(step, db=true, sdc=false, def=false, verilog=false) %}
+{% set step_id = get_step_id(step) %}
+{% if should_write_checkpoint(step) %}
 {% if db %}
-write_db {{results_dir}}/{{step_id}}.odb
+write_db {{settings.checkpoints_dir}}/{{step_id}}.odb
 {% endif %}
 {% if sdc %}
-write_sdc {{results_dir}}/{{step_id}}.sdc
+write_sdc {{settings.results_dir}}/{{step_id}}.sdc
 {% endif %}
 {% if def %}
-write_def {{results_dir}}/{{step_id}}.def
+write_def {{settings.results_dir}}/{{step_id}}.def
 {% endif %}
 {% if verilog %}
-write_verilog {{results_dir}}/{{step_id}}.v
+write_verilog {{settings.results_dir}}/{{step_id}}.v
 {% endif %}
 {% endif %}
 {% endmacro %}
